Added Canvas

+++ Assets/Scripts/Projectile.cs
@@ -4,7 +4,7 @@ public class Projectile : MonoBehaviour
 {
     public bool team;
     public bool disappearOnHit;
-    public double damageDealt;
+    public float damageDealt;
     private void OnTriggerEnter2D(Collider2D coll) {
         print("Triggered");
         if (coll.gameObject.tag == "Wall" && this.disappearOnHit) {
